Move isFieldElement to utils
