+++ services/ui-src/src/components/reports/ModalOverlayReportPageV2.tsx
@@ -20,7 +20,6 @@ import {
   EntityShape,
   ErrorVerbiage,
   FormJson,
-  isFieldElement,
   ModalOverlayReportPageShape,
   ReportShape,
   ReportStatus,
@@ -32,9 +31,10 @@ import {
   filterFormData,
   getEntriesToClear,
   getReportVerbiage,
+  isClosedInitiative,
+  isFieldElement,
   resetClearProp,
   setClearedEntriesToDefaultValue,
-  isClosedInitiative,
   useBreakpoint,
   useStore,
 } from "utils";
